requirements.txt
openai>=1.10.0
chromadb>=0.4.22
sqlalchemy>=2.0.25
pydantic>=2.5.3
python-dotenv>=1.0.1
reportlab>=4.0.9
streamlit>=1.31.0
pillow>=10.2.0
pylatexenc>=2.10
matplotlib>=3.8.0
numpy>=1.26.0
